+++ kgsm-lib/Json/KgsmJsonContext.cs
@@ -55,6 +55,8 @@ namespace TheKrystalShip.KGSM;
 [JsonSerializable(typeof(Dictionary<string, WatchdogInstancePresence>))]
 [JsonSerializable(typeof(WatchdogUpnpMapping))]
 [JsonSerializable(typeof(WatchdogUpnpList))]
+[JsonSerializable(typeof(WatchdogConsoleRun))]
+[JsonSerializable(typeof(WatchdogConsoleRun[]))]
 // --- Event envelope + every event data payload (the Type-keyed dispatch in EventService) ---
 [JsonSerializable(typeof(EventWrapper))]
 [JsonSerializable(typeof(EventCursor))]